# src/azure-ai-foundry/foundry-agentic-mesh/services/mcp-docs-server/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Docs Server
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      operationId: health
      responses:
        "200":
          description: OK
  /tools/search_docs:
    post:
      operationId: search_docs
      description: Search internal documents (via Azure AI Search) and return relevant snippets.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                top_k:
                  type: integer
                  minimum: 1
                  maximum: 20
                  default: 5
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      required: [title, snippet, source]
                      properties:
                        title: { type: string }
                        snippet: { type: string }
                        source: { type: string }
                        score: { type: number }
  /tools/create_change_request:
    post:
      operationId: create_change_request
      description: Create a change request draft artifact (JSON + Markdown).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, summary, risk_level, steps, rollback_plan]
              properties:
                title: { type: string }
                summary: { type: string }
                risk_level:
                  type: string
                  enum: [low, medium, high]
                steps:
                  type: array
                  items: { type: string }
                rollback_plan: { type: string }
                approvers:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: Draft created
          content:
            application/json:
              schema:
                type: object
                required: [id, status, location]
                properties:
                  id: { type: string }
                  status: { type: string }
                  location: { type: string }